fix converter crashes
In the converter, the AssetManager only allocated BaseAsset types, and
filled in the name. So all the other derived type members were garbage,
but some of the code during scene loading was still trying to access
those. Changed it to allocate the derived type, so everything is default
strings, vectors, etc. Also stopped creating the Lua scripts, since they
havent been loaded in anyways, just the script name