fixed model tangents, added a few models
Also made change so that you don't need to specify any default materials
on a pmodel file, it will auto fill with default materials if missing.
Reorganized the default assets into their own fastfile too.

The BVH seems to be working now? The bunny produced the same result with
and without the BVH, but super fast with the BVH. Unfortunately, the
result was wrong (a few small holes, and lots of bright specs)