added shader include cache
Time for converter to run with no assets out of date went from ~0.70
seconds to 0.03.

Other changes include:
- Adding a shader preproc cache. Once a file is included once, it's
stored in memory. Doesn't seem to make much difference at the moment,
but keeping it to test when I have more shaders
- Made all invalid filestamps (source or dependent files) mark an asset
as out of date. Previously it was only the source file
- All parsed createInfos have absolute paths now.
    - I actually don't think i want this, but easier for now