fix build issue with freetype2 again
Turns out the gitignore file was preventing some of the files from being
committed (from the build folder). I made the files in the gitignore
only apply from the root folder now, which was the intent all along for
most of them.

I also upgraded to 2.13.2 in the process, but it's unrelated to the
build fix