cmake work, and shared lib fixups
- some of the external libraries that are being used, expect normal
cmake inclusions (add_subdirectory). When you link them, it will
automatically define a compile definition to turn on the dllimport when
necessary. Since I compile the libs outside of the project though, I
need to manually define this.
- Added some missing project's source files such that they are findable
in VS file search, but not compiled against
- For both of the above changes, I shoved all the libs together into 2
big variables in cmake, for convenience