pbr + sponza scenes working again on AMD card
My 7900xt only allows 128 bytes of push constants, and I was using
something like 178 in the lit shader. Just switched to always uploading
the model and normal matrices to a buffer and read from that instead.
Not anything special or profiled at all, I just want to get it running
again