fixed fastfiles sometimes skipping rebuild
If you built a fastfile, changed an asset, rebuilt again, everything
would be fine so far. But if you then reverted the asset change and
tried to build again, it wouldn't rebuild the fastfile because no
individual assets were out of date (the previous asset still has a
valid cache entry).

- Now save out a list of all assets in each fastfile
- That list will be read the next time the converter processes that
  fastfile, and will check it to see if it needs to rebuild given the
  current scene
- No longer need to have a timestamp dependency on the scene files
  themselves

Misc changes:
- Also changed converters from shared_ptrs to new/delete
- Made the FileReadResult be non-copyable, but moveable, with a
  destructor that automatically frees the data